changed filter behaviour

+++ cogs/chat_filter.py
@@ -1,12 +1,14 @@
 from discord.ext import commands
-import os
+import os, random
 
+from asyncio import sleep
 from lib.db import db
 
 
 class Filter(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
+        self.emojis = ["🚫", "😱", "🤬", "👺", "🐒"]
 
     # only if server has chat filter enabled
     async def cog_check(self, ctx):
@@ -20,12 +22,26 @@ class Filter(commands.Cog):
             settings = db.servers.find_one(
                 {"_id": msg.guild.id}, {"chat_filter": 1, "badwords": 1, "prefix": 1}
             )
-            if settings["chat_filter"]:
+            if settings["chat_filter"] and not msg.content.startswith(
+                settings["prefix"]
+            ):
+                violations = 0
+                used_words = []
                 for word in settings["badwords"]:
-                    if word in msg.content and not msg.content.startswith(
-                        settings["prefix"]
-                    ):
-                        await msg.delete()
+                    if word in msg.content:  # msg contains banned words
+                        violations += 1
+                        used_words.append(word)
+                        await msg.add_reaction(random.choice(self.emojis))
+
+                if violations > 0:
+                    warn = await msg.channel.send(
+                        f"{msg.author.mention}! {violations} violation(s): {', '.join(used_words)}\n❗ Deleting in 5 seconds."
+                    )
+
+                    await sleep(5)
+
+                    await msg.delete()
+                    await warn.delete()
 
     @commands.group(
         aliases=["f"],
